+++ contracts/test/TickBitMapTest.sol
@@ -9,45 +9,45 @@ contract TickBitMapTest {
 
     uint256[58] public bitmap;
 
-    function isInitialized(int16 tick) external view returns (bool) {
+    function isInitialized(int32 tick) external view returns (bool) {
         return bitmap.isInitialized(tick);
     }
 
-    function getGasCostOfIsInitialized(int16 tick) external view returns (uint256) {
+    function getGasCostOfIsInitialized(int32 tick) external view returns (uint256) {
         uint256 gasBefore = gasleft();
         bitmap.isInitialized(tick);
         return gasBefore - gasleft();
     }
 
-    function flipTick(int16 tick) external {
+    function flipTick(int32 tick) external {
         bitmap.flipTick(tick);
     }
 
-    function getGasCostOfFlipTick(int16 tick) external returns (uint256) {
+    function getGasCostOfFlipTick(int32 tick) external returns (uint256) {
         uint256 gasBefore = gasleft();
         bitmap.flipTick(tick);
         return gasBefore - gasleft();
     }
 
-    function nextInitializedTickWithinOneWord(int16 tick, bool lte)
+    function nextInitializedTickWithinOneWord(int32 tick, bool lte)
         external
         view
-        returns (int16 next, bool initialized)
+        returns (int32 next, bool initialized)
     {
         return bitmap.nextInitializedTickWithinOneWord(tick, lte);
     }
 
-    function getGasCostOfNextInitializedTickWithinOneWord(int16 tick, bool lte) external view returns (uint256) {
+    function getGasCostOfNextInitializedTickWithinOneWord(int32 tick, bool lte) external view returns (uint256) {
         uint256 gasBefore = gasleft();
         bitmap.nextInitializedTickWithinOneWord(tick, lte);
         return gasBefore - gasleft();
     }
 
-    function nextInitializedTick(int16 tick, bool lte) external view returns (int16 next) {
+    function nextInitializedTick(int32 tick, bool lte) external view returns (int32 next) {
         return bitmap.nextInitializedTick(tick, lte);
     }
 
-    function getGasCostOfNextInitializedTick(int16 tick, bool lte) external view returns (uint256) {
+    function getGasCostOfNextInitializedTick(int32 tick, bool lte) external view returns (uint256) {
         uint256 gasBefore = gasleft();
         bitmap.nextInitializedTick(tick, lte);
         return gasBefore - gasleft();
